WIP: update with latest modifs - tyupes names and place, new emitter style, adapted map elites
